Optimization. Implemented Image and Color.

+++ ml-toolbox/core/ml-core-api/trunk/src/main/java/se/jtiden/ml/core/api/Hypothesis.java
@@ -1,7 +1,5 @@
 package se.jtiden.ml.core.api;
 
-import java.math.BigInteger;
-
 public interface Hypothesis extends Comparable<Hypothesis> {
-    BigInteger valueFunction();
+    double valueFunction();
 }
